requirements.txt
# ===========================================
# Runtime Dependencies (no PyTorch needed!)
# ===========================================

# API Framework
fastapi>=0.104.0,<1.0
uvicorn[standard]>=0.24.0

# LLM & Embeddings
openai>=1.6.0
voyageai>=0.3.0
langgraph>=0.0.20

# Database
supabase>=2.3.0

# ML Inference (ONNX - lightweight, ~50MB total)
onnxruntime>=1.16.0
transformers>=4.36.0,<5.0
numpy>=1.24.0

# Model Hub (download ONNX models from HuggingFace)
huggingface_hub>=0.20.0

# Auth
PyJWT[crypto]>=2.8.0

# Data validation
pydantic>=2.5.0

# Environment
python-dotenv>=1.0.0

# HTTP client (for Harmonia API calls)
requests>=2.31.0

# Language detection (FastText lid.176.bin model via HuggingFace Hub)
fasttext>=0.9.2
# Translation (Google Translate via deep-translator, no API key needed)
deep-translator>=1.11.0
